refactor(workitems): extract WorkItemContent list and editing hooks
Move the linked-session table into LinkedSessionsList plus
linkedSessionPresentation, and description draft/save and handoff
response state into hooks/. index.tsx keeps composition only.
Behaviour unchanged.
