+++ src/modules/ProjectManager/WorkItems/components/WorkItemContent/index.tsx
@@ -1,14 +1,8 @@
 import { useAtomValue } from "jotai";
-import { Bot, Pencil, Repeat, RotateCcw, Terminal } from "lucide-react";
+import { Pencil, Repeat } from "lucide-react";
 import React, { useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 
-import {
-  type WorkItemHandoff,
-  type WorkItemOriginSession,
-  projectApi,
-} from "@src/api/http/project";
-import Org2SessionIcon from "@src/assets/modelIcons/org2-session.svg";
 import Avatar from "@src/components/Avatar";
 import TabPill from "@src/components/TabPill";
 import { useWorkItemImageInsert } from "@src/hooks/project";
@@ -26,28 +20,15 @@ import {
   TimelineCardHeader,
   TimelineStack,
 } from "@src/modules/shared/components/ActivityTimeline";
-import MarkdownTextareaEditor, {
-  type MarkdownEditorMode,
-} from "@src/modules/shared/components/MarkdownTextareaEditor";
+import MarkdownTextareaEditor from "@src/modules/shared/components/MarkdownTextareaEditor";
 import MarkdownEditorModeSwitch from "@src/modules/shared/components/MarkdownTextareaEditor/ModeSwitch";
-import {
-  formatTokensShort,
-  formatUsd,
-} from "@src/modules/shared/dataSource/usageFormat";
 import {
   DetailPanelContainer,
   PanelFooter,
   ScrollTrailTarget,
-  SessionTable,
-  type SessionTableItem,
 } from "@src/modules/shared/layouts/blocks";
-import { type LinkedSession, WORK_ITEM_STATUS } from "@src/types/core/workItem";
-import {
-  formatReplayDateLabel,
-  toIntlLocaleTag,
-} from "@src/util/data/formatters/date";
+import { WORK_ITEM_STATUS } from "@src/types/core/workItem";
 
-import { STATUS_I18N_KEYS } from "../AgentWorkflow/types";
 import WorkItemContentStack from "../WorkItemContentStack";
 import WorkItemSubItems, { useWorkItemFamily } from "../WorkItemSubItems";
 import {
@@ -58,276 +39,17 @@ import {
 import CustomPropertiesSection from "./CustomPropertiesSection";
 import GitHubIssueComposer from "./GitHubIssueComposer";
 import HistoryTab from "./HistoryTab";
+import { LinkedSessionsList } from "./LinkedSessionsList";
 import OutputTab from "./OutputTab";
 import WorkItemHandoffNotice from "./WorkItemHandoffNotice";
 import WorkItemRunUsageSummary from "./WorkItemRunUsageSummary";
 import { normalizeLegacyEscapedMarkdown } from "./descriptionMarkdown";
-import { retryFailedLinkedSession } from "./discussionCommentForward";
 import { useGitHubIssueTimeline } from "./hooks/useGitHubIssueTimeline";
 import { useWorkItemContentState } from "./hooks/useWorkItemContentState";
+import { useWorkItemDescriptionEditing } from "./hooks/useWorkItemDescriptionEditing";
+import { useWorkItemHandoff } from "./hooks/useWorkItemHandoff";
 import { resolveWorkItemContentSectionPolicy } from "./presentation";
 import type { SessionTab, WorkItemContentProps } from "./types";
-
-interface LinkedSessionsListProps {
-  sessions: LinkedSession[];
-  originSession?: WorkItemOriginSession;
-  shortId?: string | null;
-  projectSlug?: string | null;
-  orgId?: string | null;
-  activeAgentSessionId?: string | null;
-  onOpenSession?: (sessionId: string) => void;
-}
-
-const LINKED_SESSION_STATUS_COLOR: Record<LinkedSession["status"], string> = {
-  running: "var(--color-primary-6)",
-  completed: "var(--color-success-6)",
-  failed: "var(--color-danger-6)",
-  cancelled: "var(--color-warning-6)",
-};
-
-const LINKED_SESSION_AGENT_LABEL: Record<LinkedSession["agent_role"], string> =
-  {
-    coding: "Coding",
-    sde: "SDE",
-    review: "Review",
-    orchestrator: "Orchestrator",
-    custom: "Custom",
-    sub_agent: "Sub-agent",
-  };
-
-function formatLinkedSessionAgentLabel(session: LinkedSession): string {
-  return (
-    session.sub_agent_name ?? LINKED_SESSION_AGENT_LABEL[session.agent_role]
-  );
-}
-
-function formatOriginSessionAgentLabel(actorId: string): string {
-  const agentId = actorId.replace(/^agent:/, "");
-  return (
-    LINKED_SESSION_AGENT_LABEL[agentId as LinkedSession["agent_role"]] ??
-    agentId
-  );
-}
-
-function renderSessionAgentIcon(
-  sessionType: "native" | "cli",
-  provider?: string
-) {
-  if (provider === "org2" || (!provider && sessionType === "native")) {
-    return (
-      <Org2SessionIcon
-        className="size-3.5 shrink-0"
-        data-agent-provider="org2"
-        aria-hidden="true"
-      />
-    );
-  }
-
-  const AgentIcon = sessionType === "cli" ? Terminal : Bot;
-  return (
-    <AgentIcon
-      size={14}
-      strokeWidth={1.75}
-      className="shrink-0 text-text-3"
-      aria-hidden="true"
-    />
-  );
-}
-
-function getLinkedSessionTitle(session: LinkedSession): string {
-  if (session.result_preview) return session.result_preview;
-  if (session.sub_agent_name) return session.sub_agent_name;
-  return session.session_id;
-}
-
-const LinkedSessionsList: React.FC<LinkedSessionsListProps> = ({
-  sessions,
-  originSession,
-  shortId,
-  projectSlug,
-  orgId,
-  activeAgentSessionId,
-  onOpenSession,
-}) => {
-  const { t, i18n } = useTranslation(["projects", "common"]);
-  const dateTimeLabelOptions = useMemo(
-    () => ({
-      todayLabel: t("common:relativeDate.today"),
-      yesterdayLabel: t("common:relativeDate.yesterday"),
-      locale: toIntlLocaleTag(i18n.resolvedLanguage),
-    }),
-    [i18n.resolvedLanguage, t]
-  );
-  const tableItems = useMemo<SessionTableItem[]>(() => {
-    if (sessions.length === 0 && !originSession) {
-      return [
-        {
-          id: "work-item-linked-sessions-empty",
-          title: t("workItems.sessions.emptyOverview"),
-          statusLabel: "—",
-          disabled: true,
-          testId: "work-item-linked-sessions-empty-row",
-        },
-      ];
-    }
-
-    const executionItems = sessions.map((session) => {
-      const statusLabelKey = STATUS_I18N_KEYS[session.status];
-      const statusLabel = statusLabelKey ? t(statusLabelKey) : session.status;
-
-      return {
-        id: session.session_id,
-        title: getLinkedSessionTitle(session),
-        description:
-          session.result_preview &&
-          session.result_preview !== session.session_id
-            ? session.session_id
-            : undefined,
-        statusLabel,
-        statusColor: LINKED_SESSION_STATUS_COLOR[session.status],
-        agentIcon: renderSessionAgentIcon(session.session_type),
-        agentLabel: formatLinkedSessionAgentLabel(session),
-        modelLabel: session.session_type,
-        workspaceLabel: session.parent_session_id,
-        workspaceTitle: session.parent_session_id,
-        startedLabel: formatReplayDateLabel(session.started_at, {
-          ...dateTimeLabelOptions,
-          withSeconds: false,
-          monthStyle: "short",
-        }),
-        lastUpdatedLabel: formatReplayDateLabel(
-          session.completed_at ?? session.started_at,
-          {
-            ...dateTimeLabelOptions,
-            withSeconds: false,
-            monthStyle: "short",
-          }
-        ),
-        tokensLabel:
-          session.total_tokens > 0
-            ? formatTokensShort(session.total_tokens)
-            : undefined,
-        tokensValue:
-          session.total_tokens > 0 ? session.total_tokens : undefined,
-        active: session.session_id === activeAgentSessionId,
-        testId: `work-item-linked-session-${session.session_id}`,
-        rowAction:
-          session.status === "failed" && shortId ? (
-            <button
-              type="button"
-              className="flex cursor-pointer items-center gap-1 rounded-md px-1.5 py-1 text-[11px] text-text-3 transition-colors hover:bg-fill-2 hover:text-text-1"
-              onClick={() => {
-                retryFailedLinkedSession({
-                  projectSlug,
-                  orgId,
-                  shortId,
-                  sessionId: session.session_id,
-                });
-                onOpenSession?.(session.session_id);
-              }}
-              aria-label={t("workItems.sessions.retry", {
-                defaultValue: "Retry",
-              })}
-              data-testid={`work-item-session-retry-${session.session_id}`}
-            >
-              <RotateCcw size={12} />
-              {t("workItems.sessions.retry", { defaultValue: "Retry" })}
-            </button>
-          ) : undefined,
-      };
-    });
-    if (
-      !originSession ||
-      sessions.some(
-        (session) => session.session_id === originSession.session_id
-      )
-    ) {
-      return executionItems;
-    }
-    return [
-      {
-        id: originSession.session_id,
-        title: t("workItems.sessions.originTitle", {
-          defaultValue: "Creation session",
-        }),
-        description: originSession.session_id,
-        statusLabel: t("workItems.sessions.originStatus", {
-          defaultValue: "Created this item",
-        }),
-        statusColor: "var(--color-primary-6)",
-        agentIcon: renderSessionAgentIcon(
-          originSession.session_type,
-          originSession.provider
-        ),
-        agentLabel: formatOriginSessionAgentLabel(originSession.actor_id),
-        modelLabel: originSession.session_type,
-        startedLabel: formatReplayDateLabel(originSession.captured_at, {
-          ...dateTimeLabelOptions,
-          withSeconds: false,
-          monthStyle: "short",
-        }),
-        lastUpdatedLabel: formatReplayDateLabel(originSession.captured_at, {
-          ...dateTimeLabelOptions,
-          withSeconds: false,
-          monthStyle: "short",
-        }),
-        active: originSession.session_id === activeAgentSessionId,
-        disabled: originSession.provider !== "org2",
-        testId: `work-item-origin-session-${originSession.session_id}`,
-      },
-      ...executionItems,
-    ];
-  }, [
-    activeAgentSessionId,
-    dateTimeLabelOptions,
-    onOpenSession,
-    originSession,
-    orgId,
-    projectSlug,
-    sessions,
-    shortId,
-    t,
-  ]);
-
-  const totalTokens = sessions.reduce(
-    (sum, session) => sum + (session.total_tokens || 0),
-    0
-  );
-  const totalCost = sessions.reduce(
-    (sum, session) => sum + (session.cost_usd || 0),
-    0
-  );
-
-  return (
-    <div data-testid="work-item-linked-sessions">
-      {sessions.length > 0 && (
-        <div
-          className="mb-1 flex items-center gap-3 px-1 text-[11px] text-text-4"
-          data-testid="work-item-usage-summary"
-        >
-          <span>
-            {t("workItems.sessions.runsCount", {
-              defaultValue: "{{count}} runs",
-              count: sessions.length,
-            })}
-          </span>
-          {totalTokens > 0 && (
-            <span>{formatTokensShort(totalTokens)} tokens</span>
-          )}
-          {totalCost > 0 && <span>{formatUsd(totalCost, 2)}</span>}
-        </div>
-      )}
-      <SessionTable
-        items={tableItems}
-        onSelect={(item) => onOpenSession?.(item.id)}
-        surfaceVariant="default"
-        bodySurface="pane"
-        headerBorder={false}
-        maxHeight={360}
-      />
-    </div>
-  );
-};
 
 const WorkItemContent: React.FC<WorkItemContentProps> = ({
   workItem,
@@ -438,18 +160,24 @@ const WorkItemContent: React.FC<WorkItemContentProps> = ({
     githubIssueTimeline?.items ?? loadedGitHubTimeline.timeline;
   const githubTimelineLoading =
     githubIssueTimeline?.loading ?? loadedGitHubTimeline.timelineLoading;
-  const [descriptionDraftState, setDescriptionDraftState] = useState<{
-    workItemId: string;
-    base: string;
-    value: string;
-  } | null>(null);
-  const [descriptionEditWorkItemId, setDescriptionEditWorkItemId] = useState<
-    string | null
-  >(null);
-  const [descriptionSaveErrorWorkItemId, setDescriptionSaveErrorWorkItemId] =
-    useState<string | null>(null);
-  const [descriptionEditorMode, setDescriptionEditorMode] =
-    useState<MarkdownEditorMode>("write");
+  const {
+    descriptionDraft,
+    descriptionHasChanges,
+    descriptionEditWorkItemId,
+    descriptionSaveErrorWorkItemId,
+    descriptionEditorMode,
+    setDescriptionEditorMode,
+    handleDescriptionDraftChange,
+    handleCancelDescription,
+    handleSaveDescription,
+    beginDescriptionEdit,
+  } = useWorkItemDescriptionEditing({
+    workItemId: workItem.session_id,
+    displayedDescription,
+    isGitHubWorkItem,
+    githubIssueInteraction,
+    onCommitDescription: handleDescriptionChange,
+  });
   const [threadViewSelection, setThreadViewSelection] = useState<{
     workItemId: string;
     view: WorkItemThreadView;
@@ -457,18 +185,6 @@ const WorkItemContent: React.FC<WorkItemContentProps> = ({
     workItemId: workItem.session_id,
     view: "overview",
   });
-  const currentDescriptionDraft =
-    descriptionDraftState?.workItemId === workItem.session_id
-      ? descriptionDraftState
-      : null;
-  const descriptionHasChanges = Boolean(
-    currentDescriptionDraft &&
-    currentDescriptionDraft.value !== currentDescriptionDraft.base
-  );
-  const descriptionDraft =
-    currentDescriptionDraft && descriptionHasChanges
-      ? currentDescriptionDraft.value
-      : displayedDescription;
   const sectionPolicy = resolveWorkItemContentSectionPolicy(
     presentation,
     Boolean(workItem.proofOfWork)
@@ -480,78 +196,24 @@ const WorkItemContent: React.FC<WorkItemContentProps> = ({
       : "overview";
   const isEditingThreadDescription =
     isThread && descriptionEditWorkItemId === workItem.session_id;
-  const [handoffOverride, setHandoffOverride] = useState<{
-    workItemId: string;
-    value: WorkItemHandoff;
-  } | null>(null);
-  const [respondingHandoff, setRespondingHandoff] = useState<
-    "accept" | "return" | null
-  >(null);
-  const [handoffError, setHandoffError] = useState<string | null>(null);
-  const handoff =
-    handoffOverride?.workItemId === workItem.session_id &&
-    handoffOverride.value.id === workItem.handoff?.id
-      ? handoffOverride.value
-      : workItem.handoff;
-  const canRespondToHandoff = Boolean(
-    handoff &&
-    shortId &&
-    (onTransitionHandoff || projectSlug) &&
-    currentUserMemberIds.has(handoff.recipientMemberId) &&
-    handoff.status === "pending"
-  );
-  const handoffRecipientName = handoff
-    ? teamMembers.find((member) => member.id === handoff.recipientMemberId)
-        ?.name
-    : undefined;
-  const handoffResponseUnavailableReason =
-    handoff?.status === "pending" && currentUser.id === "system"
-      ? t("common:teamInbox.handoff.identityUnavailable")
-      : undefined;
-
-  const respondToHandoff = (action: "accept" | "return", note?: string) => {
-    if (
-      !handoff ||
-      !shortId ||
-      (!onTransitionHandoff && !projectSlug) ||
-      respondingHandoff ||
-      !currentUserMemberIds.has(handoff.recipientMemberId)
-    ) {
-      return;
-    }
-    setRespondingHandoff(action);
-    setHandoffError(null);
-    const transition = {
-      handoffId: handoff.id,
-      action,
-      actor: {
-        id: handoff.recipientMemberId,
-        name: handoffRecipientName || handoff.recipientName || currentUser.name,
-      },
-      note,
-    } as const;
-    const request = onTransitionHandoff
-      ? onTransitionHandoff(transition)
-      : projectApi.transitionWorkItemHandoff(projectSlug!, shortId, transition);
-    void request
-      .then((result) => {
-        const nextHandoff =
-          "frontmatter" in result ? result.frontmatter.handoff : result.handoff;
-        if (nextHandoff) {
-          setHandoffOverride({
-            workItemId: workItem.session_id,
-            value: nextHandoff,
-          });
-        }
-        onRefreshWorkflow?.();
-      })
-      .catch(() => {
-        setHandoffError(t("common:teamInbox.handoff.responseError"));
-      })
-      .finally(() => {
-        setRespondingHandoff(null);
-      });
-  };
+  const {
+    handoff,
+    canRespondToHandoff,
+    handoffError,
+    handoffResponseUnavailableReason,
+    respondingHandoff,
+    respondToHandoff,
+  } = useWorkItemHandoff({
+    workItem,
+    shortId,
+    projectSlug,
+    onTransitionHandoff,
+    onRefreshWorkflow,
+    currentUser,
+    currentUserMemberIds,
+    teamMembers,
+    t,
+  });
   const handoffNotice = handoff ? (
     <WorkItemHandoffNotice
       handoff={handoff}
@@ -564,60 +226,12 @@ const WorkItemContent: React.FC<WorkItemContentProps> = ({
     />
   ) : null;
 
-  const handleDescriptionDraftChange = (markdown: string) => {
-    setDescriptionSaveErrorWorkItemId(null);
-    setDescriptionDraftState((current) => {
-      if (current?.workItemId === workItem.session_id) {
-        return { ...current, value: markdown };
-      }
-      return {
-        workItemId: workItem.session_id,
-        base: displayedDescription,
-        value: markdown,
-      };
-    });
-  };
-
-  const handleCancelDescription = () => {
-    setDescriptionDraftState(null);
-    setDescriptionEditWorkItemId(null);
-    setDescriptionSaveErrorWorkItemId(null);
-  };
-
-  const handleSaveDescription = async () => {
-    if (!descriptionHasChanges) return;
-    if (isGitHubWorkItem) {
-      if (
-        !githubIssueInteraction?.canEditBody ||
-        githubIssueInteraction.updatingBody
-      ) {
-        return;
-      }
-      try {
-        await githubIssueInteraction.onUpdateBody(descriptionDraft);
-        setDescriptionDraftState(null);
-        setDescriptionEditWorkItemId(null);
-        setDescriptionSaveErrorWorkItemId(null);
-      } catch {
-        setDescriptionSaveErrorWorkItemId(workItem.session_id);
-      }
-      return;
-    }
-    handleDescriptionChange(descriptionDraft);
-    setDescriptionDraftState(null);
-    setDescriptionEditWorkItemId(null);
-    setDescriptionSaveErrorWorkItemId(null);
-  };
-
   const descriptionActions =
     isThread && canEditDescription && !isEditingThreadDescription ? (
       <ActivityHeaderActionButton
         icon={<Pencil size={12} aria-hidden />}
         label={t("common:actions.edit")}
-        onClick={() => {
-          setDescriptionSaveErrorWorkItemId(null);
-          setDescriptionEditWorkItemId(workItem.session_id);
-        }}
+        onClick={beginDescriptionEdit}
         data-testid="work-item-description-edit"
       />
     ) : null;
